requirement.txt
flask
pandas
numpy
matplotlib
requests
connexion
Flask-SQLAlchemy
config
flask_marshmallow
flask_cors
Flask-Script
Flask-Migrate
marshmallow-sqlalchemy